task2_grocery_store/requirements.txt
Django==5.1.3
django-imagekit==5.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
Pillow==11.0.0